# Performance Improvement: replace null with sentinel value