This is an accidental discovery from writing [parinfer-jvm].

The main result object has a lot of "integer or null" values. In order to get around the JVM type system I created a `SENTINEL_NULL` integer to represent null. ([Sentinel value])

Decided to add it to `parinfer.js` to see what it did to performance and saw some impressive results:
[performance with regular null](https://gist.github.com/oakmac/ede280269cfd9660f051)
[performance with SENTINEL_NULL](https://gist.github.com/oakmac/b4200f2cf98f856af32b)

Interestingly, this change had no performance improvement to [parinfer.py]. If anything, it was a touch slower.

My guess is this has something to do with how JS engines do type coercion.

Here is a summary of the performance result averages:

|                                              | using null | using SENTINEL_NULL |
|----------------------------------------------|------------|---------------------|
| long_map_with_strings: indent                | 10.2958ms  | 9.0249ms            |
| long_map_with_string:paren                   | 8.4173ms   | 4.09438ms           |
|                                              |            |                     |
| really_long_file: indent                     | 24.1756ms  | 12.6804ms           |
| really_long_file: paren                      | 32.1068ms  | 23.4434ms           |
|                                              |            |                     |
| really_long_file_with_unclosed_paren: indent | 33.4156ms  | 9.9237ms            |
| really_long_file_with_unclosed_paren: paren  | 40.7618ms  | 22.5091ms           |
|                                              |            |                     |
| really_long_file_with_unclosed_quote: indent | 40.9007ms  | 9.0063ms            |
| really_long_file_with_unclosed_quote: paren  | 41.7376ms  | 23.3028ms           |

[parinfer-jvm]:https://github.com/oakmac/parinfer-jvm
[Sentinel value]:https://en.wikipedia.org/wiki/Sentinel_value
[parinfer.py]:https://github.com/oakmac/parinfer.py